WORKS!

+++ Karatsuba/intArr.h
@@ -14,9 +14,8 @@ namespace Mult
         void insert(int place, int value);
         void pushback(int value);
         int getActualSize()const;
-        int getSize() { return intArray.size(); }
+        int getSize() { return (int)intArray.size(); }
         int get(int place)const { return intArray[place]; }
-        //int getback() { return intArray.back(); }
         void printArr();
         void resize();
         void shiftLeft(int toshift);
